update lspconfig setup
reworked it using kickstart as an example. was never happy with my
original one. while it was nice and home grown, it wasn't very idiomatic
and i'm pretty sure it was setting up language servers twice. never
caused any issues but this seems to work fine and is a more common
configuration anyway