add gitsigns and satellite
gitsigns is replacing mini.diff. it has more features that are sometimes
useful (like git blame) and integrates with satellite.

im using satellite instead of mini.map because i prefer the way it
looks. maybe a skill issue on the part of configuring the display of
mini.map, but at that point i was already integrating gitsigns (since
mini.map would have required it anyway) and then gitsigns and satellite
are by the same author so whatever